Rename ProjectInit into ProjectInfo
