+++ asreview/webapp/src/Projects.js
@@ -9,7 +9,7 @@ import ProjectCard from "./ProjectCard";
 
 import { ImportDialog, QuickTourDialog } from "./Components";
 
-import { ProjectInit } from "./PreReviewComponents";
+import { ProjectInfo } from "./PreReviewComponents";
 
 import { ProjectAPI } from "./api/index.js";
 
@@ -161,7 +161,7 @@ const Projects = (props) => {
       </Container>
 
       {open.newProject && (
-        <ProjectInit
+        <ProjectInfo
           handleAppState={props.handleAppState}
           open={open.newProject}
           onClose={handleCloseNewProject}
